Added support for multiple iterators. Fixed a couple of bugs. Fixed the size of driver messages to 2 bytes.

+++ lang/cs/Org.Apache.REEF.Network/Elastic/Operators/Logical/Impl/DefaultAggregationRing.cs
@@ -94,7 +94,7 @@ namespace Org.Apache.REEF.Network.Elastic.Operators.Logical.Impl
             {
                 case TaskMessageType.JoinTopology:
                     {
-                        var operatorId = BitConverter.ToInt32(message.Message, sizeof(ushort));
+                        var operatorId = BitConverter.ToInt16(message.Message, sizeof(ushort));
 
                         if (operatorId != _id)
                         {
@@ -112,7 +112,7 @@ namespace Org.Apache.REEF.Network.Elastic.Operators.Logical.Impl
                     }
                 case TaskMessageType.TopologyUpdateRequest:
                     {
-                        var operatorId = BitConverter.ToInt32(message.Message, sizeof(ushort));
+                        var operatorId = BitConverter.ToInt16(message.Message, sizeof(ushort));
 
                         if (operatorId != _id)
                         {
@@ -136,7 +136,7 @@ namespace Org.Apache.REEF.Network.Elastic.Operators.Logical.Impl
                     }
                 case TaskMessageType.NextDataRequest:
                     {
-                        var iteration = BitConverter.ToInt32(message.Message, sizeof(ushort));
+                        var iteration = BitConverter.ToInt16(message.Message, sizeof(ushort));
                         LOGGER.Log(Level.Info, "Received next data request from {0} for iteration {1}", message.TaskId, iteration);
 
                         RingTopology.ResumeRing(ref returnMessages, message.TaskId, iteration);
